♻️🎪 Extract loader from route
