+++ apps/show/src/routes/_exhibitor.exposants.$token._config._index/loader.server.ts
@@ -1,0 +1,27 @@
+import { services } from "#core/services.server.js";
+import { RouteParamsSchema } from "#exhibitors/route-params";
+import { safeParseRouteParam } from "@animeaux/zod-utils";
+import type { LoaderFunctionArgs } from "@remix-run/node";
+
+export async function loader({ params }: LoaderFunctionArgs) {
+  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
+
+  const exhibitor = await services.exhibitor.getByToken(routeParams.token, {
+    select: {
+      token: true,
+      sponsorship: { select: { category: true } },
+      documentStatus: true,
+      documentStatusMessage: true,
+      dogsConfigurationStatus: true,
+      name: true,
+      publicProfileStatus: true,
+      descriptionStatus: true,
+      onStandAnimationsStatus: true,
+      standNumber: true,
+      standConfigurationStatus: true,
+      invoices: { select: { status: true } },
+    },
+  });
+
+  return { exhibitor };
+}
