+++ apps/show/src/routes/_exhibitor.exposants.$token._config._index/route.tsx
@@ -2,11 +2,8 @@ import { getErrorTitle } from "#core/data-display/error-page";
 import { FormLayout } from "#core/layout/form-layout";
 import { createSocialMeta } from "#core/meta";
 import { getPageTitle } from "#core/page-title";
-import { services } from "#core/services.server.js";
-import { RouteParamsSchema } from "#exhibitors/route-params";
-import { safeParseRouteParam } from "@animeaux/zod-utils";
-import type { LoaderFunctionArgs } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
+import type { loader } from "./loader.server";
 import { SectionAwaitingValidation } from "./section-awaiting-validation";
 import { SectionHelper } from "./section-helper";
 import { SectionPayment } from "./section-payment";
@@ -15,28 +12,7 @@ import { SectionStandNumber } from "./section-stand-number";
 import { SectionToComplete } from "./section-to-complete";
 import { SectionValidated } from "./section-validated";
 
-export async function loader({ params }: LoaderFunctionArgs) {
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const exhibitor = await services.exhibitor.getByToken(routeParams.token, {
-    select: {
-      token: true,
-      sponsorship: { select: { category: true } },
-      documentStatus: true,
-      documentStatusMessage: true,
-      dogsConfigurationStatus: true,
-      name: true,
-      publicProfileStatus: true,
-      descriptionStatus: true,
-      onStandAnimationsStatus: true,
-      standNumber: true,
-      standConfigurationStatus: true,
-      invoices: { select: { status: true } },
-    },
-  });
-
-  return { exhibitor };
-}
+export { loader } from "./loader.server";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return createSocialMeta({
