+++ apps/show/src/routes/_exhibitor.exposants.$token._config._index/section-awaiting-validation.tsx
@@ -1,7 +1,7 @@
 import { FormLayout } from "#core/layout/form-layout";
 import { ShowExhibitorStatus } from "@prisma/client";
 import { useLoaderData } from "@remix-run/react";
-import type { loader } from "./route";
+import type { loader } from "./loader.server";
 import {
   TaskItemDescription,
   TaskItemDocument,
