+++ apps/show/src/routes/_exhibitor.exposants.$token._config._index/section-payment.tsx
@@ -3,7 +3,7 @@ import { HelperCard } from "#core/layout/helper-card";
 import { Routes } from "#core/navigation.js";
 import { InvoiceStatus } from "#invoice/status.js";
 import { Link, useLoaderData } from "@remix-run/react";
-import type { loader } from "./route";
+import type { loader } from "./loader.server";
 
 export function SectionPayment() {
   const { exhibitor } = useLoaderData<typeof loader>();
