+++ apps/show/src/routes/_exhibitor.exposants.$token._config._index/section-stand-number.tsx
@@ -1,6 +1,6 @@
 import { HelperCard } from "#core/layout/helper-card";
 import { useLoaderData } from "@remix-run/react";
-import type { loader } from "./route";
+import type { loader } from "./loader.server";
 
 export function SectionStandNumber() {
   const { exhibitor } = useLoaderData<typeof loader>();
